TraceAnalyzer: Add code to identify ghost tasks
With perf, you can get a sched_switch event, where the pid of the event
differs from oldpid. In these cases the event pid is considered a ghost
pid because it is apparently never scheduled. It just appears as the pid
of the sched_switch, wakeup and waking events of certain tasks.

I have named these pids/tasks "ghost alias tasks/pids". They appear to be
a sort of alias for certain tasks. This has been seen in a container case,
when perf is run inside a container, so it's possibly pid name space
related.

The main reason why we want to identify these tasks is that when we are
using the buttons to find wakeup and waking events, we want to select the
task that was executing, and since we get the pid from the event, we get
the ghost pid and we need to be able to find the real task when we have
only the ghost pid. For this reason, a new function, findRealTask() is
added to the TraceAnalyzer class.

Also, if the user selects a ghost in the task selector and push the button
to display a unified graph, then we are going to display the real task,
because the ghost has no scheduling events.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>