AbstractTask: Make the events member static
I don't like these static members because they become a sort of global
variable. However, I feel that since we already have those time related
static variables, we can have one for the events list as well. Currently,
all those per object events pointers are currently pointing to the same
events list, because there is only one in the whole program.

The savings in processing time and memory are not that great but I feel
that in a way it make the program more robust because we don't need to
remember to set the events member every time we create a new task. Also, at
the same time, the member is changed to const, which will make it a little
bit safer, since the AbstractTask code doesn't need to modify events.

In the future, if we wish to allow the program to have several traces
opened simultaneously, we would need to change all these static members in
AbstractTask, so that these data would be recorded in a per trace storage
somewhere, instead of being global.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>