TaskSelectDialog: Display ghost tasks differently
This change will cause the TaskSelectDialog to display ghost tasks
differently: they will be displayed with cyan background and red
foreground.

It's of interest to the user to know whether a task is a ghost task or not,
in order to decrease the confusion when the user uses the TaskSelectDialog
class to add a unified graph, because in that case the unified graph will
be of another task.

Note that the nothing of the implementation is in TaskSelectDialog, all
code for this change is in the model classes used by TaskSelectDialog.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>