MainWindow: Let the user add a ghost task to the legend
The action when the user tries to add a ghost task to the legend should
be to add the task that the ghost is an alias for. When it comes to event
filtering, we leave it as is, since there I feel that the logical
thing is to show only those events that contain the ghost pid.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>