Revert "qcustomplot: Add the SPDX license identifiers"
This reverts commit 97dc25b9149c22b490e907f2eff52f0bc9fdb944.

This commit is everted for legal reasons. When I created this commit, I
used the wrong Author information by mistake, because my git was
misconfigured. The author name indicated that this was done for my previous
employer, although it was done privately and I had left that employment 17
days earlier.