TraceAnalyzer: Check value from sysconf(_SC_PAGESIZE)
The value from sysconf(_SC_PAGESIZE) may be negative. For that reason
alone, we need to make sure that it's positive. Because we are also
calculating WRITE_BUFFER_LIMIT with a negative offset from
WRITE_BUFFER_SIZE, we need to make sure that it's large enough.

This is something that that was found with Coverity. This change
should fix CID 1530657 and CID 1530656.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>