StatsLimitedModel: Make it a subclass of StatsModel
Currently, both StatsModel and StatsLimitedModel are derived directly from
AbstractTaskModel.

The problem with this is that since both classes adds very similar code to
AbstractTaskModel, it means that there is copy pasted code with a few
modifications.

This change changes the inheritance structure so that the inhertitance
looks like this:

  AbstractTaskModel => StatsModel => StatsLimitedModel

The common code is in the non-virtual functions of StatsModel and the
code differences are moved to 7 new virtual functions that are
reimplemented in StatsLimitedModel.

This will remove the copy pasted code in StataLimitedModel, indeed this
change removes more lines of code than what it adds.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>