AbstractTaskModel: Move statistics code to StatsModel
Since we just refactored the inheritance of the StatsModel and
StatsLimitedModel classes, it means that we now can move statiscs related
code into StatsModel without having to duplicate the code in
StatsLimitedModel.

The only reason that the code was added to the AbstractTaskModel class was
that we did not want to duplicate it in StatsLimitedModel. However, now
when StatsLimitedModel inherits StatsModel, that problem no longer exist.

Hence, let's move the exportStats() and related code to StatsModel.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>