StatsModel::rowToPct(): Do not add leading spaces for lower values
This functions used to add one or two spaces before the percentage, if the
percentages were lower than 100 or 10 respectively. This doesn't really
make the statistics views look any better and it looks really bad when this
function is used to generate exported files.

For this reason, let's remove these spaces.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>