MainWindow: Make the behavior with file formats and suffixes more uniform
MainWindow has various functions that let the user save some data to a
file; in these cases we should add the correct suffix if the user has not
already done so.

Furthermore, we will allow the user to select the desired format in the
combo box in QFileDialog::getSaveFileName(), in those case where we permit
different file formats.

The user will be able to select the desired format also in the case where
the widget from which the request comes already has a way of selecting the
desired format. This is the case with exporting latencies and statistics
because these can be exported as wither ASCII or CSV, with an option to
select the a preference in the widget where the export button is.

In these cases, the format will be the default in the combobox
QFileDialog::getSaveFileName() but the user will be able to override it by
selecting a different value in the combo box.

This change also changes the behavior of the screenshot button. Previously,
the format was selected by the suffix typed by the user, with png as
default. With this change, the format is selected by the combo box in
QFileDialog::getSaveFileName() and based on this a suitable suffix is
enforced.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>