MainWindow: Use desired format if user has typed a suffix
If the user has taken the trouble to type in a file name suffix, like .csv,
or .txt, then that will be given the highest priority in selecting the file
format when saving/exporting a file.

If no suffix is given by the user, then the behavior will be as before,
i.e. the format chosen in the combo box in the
QFileDialog::getSaveFileName() dialog will be followed.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>