Make perf sched_wakeup argument parsing work with newer perf
It was disovered when testing with perf from the 6.3.y kernels that
traceshark no longer can process the wakeup event because the format
has changed again. The problem likely already exist in the 6.2.y kernel
series but this has not been tested.

For this reason we need to modify these functions. With this modification,
we should be able to deal with the following five argument formats:

<PNAME>:<PID> [<PRIO>] CPU:<CPU>
<PNAME>:<PID> [<PRIO>] success=1 CPU:<CPU>
<PNAME>:<PID> [<PRIO>]<CANT FIND FIELD success> CPU:<CPU>
comm=<PNAME> pid=<PID> prio=<PRIO> target_cpu=<CPU>
comm=<PNAME> pid=<PID> prio=<PRIO> success=1 target_cpu=<CPU>

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>