analyzer: Generate a name for ghost tasks
Ghost tasks will end up without a task name, beause of the previous commit:
4281a38 TraceAnalyzer: Improve the process name detection

This commit removed the usage of event names to detemin the name of task.
With a ghost task, we will not get any other name, because it is never
scheduled, will never be woken up, etc, so we will never get the name
from the event arguments of sched_switch or sched_wakeup events.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>