New feature: save manually selected colors
If the user takes the trouble to improve the color scheme of tasks, then it
is desirable to save these, so that they are retained the next time the
same file is opened.

The file will be saved to a file that has the same name as the trace file,
except that the suffix will be .tssetting, so that if a trace file is
named perf.asc, the state file will be names perf.tssetting.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>