bugfix: Fix broken process name detection
The following commit broke process name detection:
4281a38 TraceAnalyzer: Improve the process name detection

The problem with not using the event process name of the switch events is
that the process name is not checked often enough. For this reason we now
want to now always use the names from the arguments of the switch events.
Previously, we only did that when we had a new task but that is no longer
enough.

We do not want to go back to using the event process name of the switch
events, since that name is sometimes garbled, Instead, let's use the name
from the parameters/arguments of the switch events.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>