TaskSelectDialog: Make it possible to export from the regular dialog
Previously, it was only possible to export from the statistics dialogs.
However, I have noticed that in some cases, it would be nice to be able to
export just a list of tasks, their PIDs, and ghost status.

For this reason, let's enable the exporting also when the dialog is used
just as a plain task list, this is the list you get when you push the Task
Select Dialog button.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>