TaskSelectDialog: prevent redundant requests
When we request a task being added to the legend, or request a task graph
for a set of task, there will be several request for each task. This
happens because the taskView->selectedIndexes() will return a list of each
selected cell, and there are multiple cells per row, one for every column,
so it means that we will have one request for every column.

This is not so bad, since a task graph will not be added if there already
is one for the task, likewise if a task is already added to the legend, it
will not be added twice. However, it's a bit of a silly overhead, therefore
we will use a QMap to check that we only issue one signal per task.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>