MainWindow: Disable the reset color action when closing a trace
The action (button and menu item) for reseting the task colors remain
enabled, even if the trace is closed. This is a bug, when the trace is
closed, there are no tasks and no colors that would be applicable.

This change adds the ncessary call to disable the action when the trace is
closed.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>