Add support for building with Qt6
Qt6 is the wave of the future so it's important to support it. At the same
time, fix the build for Qt4. It's not recommended to use Qt4 but since
there doesn't seem to be a strong reason to drop support for it, let's
add the necessary code to build with Qt4. However, I expect that I will
drop support for Qt4 soon because it's now getting very old.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>