MainWindow: Improve error message when state file cannot be saved
Add the filename to the message when the state file cannot be saved. This
may make it more understandable to the user, because the filename includes
the full path. The reason for not being able to write the state file may
be that the whole directory has been deleted by the user. In this case,
the user probably knows this and can understand better that the statefile
is saved in the same directory as the trace file and can thus understand
the problem better.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>