TraceAnalyzer::processSwitchEvent(): Initialize stack variables
We better initialize as many stack variables as possible in this function.
I believe that all the warnings from g++ 12 about uninitialized use are
bogus. However, I see that if the function is quite fragile as is and it
would be quite easy to modify so that uninitialized use happens.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>