EventsWidget:handleDoubleClick(): Fix bug with incorrect column in event
The previous commit: 3cd388e ("EventsModel: Display ftrace event flags if
 available"), introduced a bug because EventsWidget was not updated
 accordingly.

The eventDoubleClicked has incorrect column information if and only if
there was no flags field, which is the case in all perf traces.

We solve this by adding the necessary translation into the column_to_int()
and int_to_column() functions of the EventsModel class. These functions
add and subtract as necessary so that whenever we convert from one type
to the other, we automagically get the corrected value. The columnt_t
values should generally not be corrected, since they are compared (e.g. in
switch statements to the enum constants), while the column as an integer
refers to position, and this value need to be corrected when one field,
i.e. COLUMN_FLAGS, is not present.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>