traceshark.pro: Add compiler optimization options for RK3588
This change adds commented out compiler options for RK3588 based boards,
like the Orange Pi 5 Plus. Also, change the comment about this being the
Rasperrby Pi section to ARM64 section, since it's now contain several
boards that are not Raspberry Pi.

We add a second entry for RK3588 because clang appears to not support the
option cortex-a76.cortex-a55, which gcc uses to let the user specify a
big.LITTLE, like the one in RK3588, so for clang we use cortex-a55; there
appears to be little performance difference between optimizing for
cortex-a76 vs cortex-a55. It is difficult to establish whether the
difference is statistically significant.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>